finished first 3 labs of lesson 08, ex04 still do

+++ PyLesson_08/PyLab_08_ex03_Lucky7s.py
@@ -2,10 +2,10 @@ number = int(input("Please enter an integer:"))
 
 def luck(number):
     if number > 0:
-        if number%10 == 7:
-            return 1 + luck(number/10)
+        if number % 10 == 7:
+            return 1 + luck(int(number/10))
         else:
-            return 0 + luck(number/10)
+            return 0 + luck(int(number/10))
     else:
         return 0
